move `ChannelUpdater` into its own file
