+++ src/aics-image-viewer/components/App/index.tsx
@@ -20,6 +20,7 @@ import {
   GlobalViewerSettings,
   ViewerSettingUpdater,
   ViewerSettingChangeHandlers,
+  UseImageEffectType,
 } from "./types";
 import { controlPointsToLut } from "../../shared/utils/controlPointsToLut";
 import {
@@ -51,6 +52,7 @@ import {
   SCALE_BAR_MARGIN_DEFAULT,
 } from "../../shared/constants";
 
+import ChannelUpdater from "./ChannelUpdater";
 import ControlPanel from "../ControlPanel";
 import Toolbar from "../Toolbar";
 import CellViewerCanvasWrapper from "../CellViewerCanvasWrapper";
@@ -186,88 +188,6 @@ function initializeLut(aimg: Volume, channelIndex: number, channelSettings?: Vie
   aimg.setLut(channelIndex, lutObject.lut);
   return newControlPoints;
 }
-
-/** `typeof useEffect`, but the effect handler takes a `Volume` as an argument */
-type UseImageEffectType = (effect: (image: Volume) => void | (() => void), deps: ReadonlyArray<any>) => void;
-
-interface ChannelUpdaterProps {
-  index: number;
-  state: ChannelState;
-  view3d: View3d;
-  image: Volume | null;
-  imageLoaded: boolean;
-}
-
-// Each channel gets a renderless updater component to sync state to viewer
-const ChannelUpdater: React.FC<ChannelUpdaterProps> = ({ index, state, view3d, image, imageLoaded }) => {
-  const { volumeEnabled, isosurfaceEnabled, isovalue, colorizeEnabled, colorizeAlpha, opacity, color, controlPoints } =
-    state;
-
-  // Effects to update channel settings should check if image is present and loaded first
-  const useImageEffect: UseImageEffectType = (effect, deps) => {
-    useEffect(() => {
-      if (image && imageLoaded) {
-        return effect(image);
-      }
-    }, [...deps, image, imageLoaded]);
-  };
-
-  useImageEffect(
-    (image) => {
-      view3d.setVolumeChannelEnabled(image, index, volumeEnabled);
-      view3d.updateLuts(image);
-    },
-    [volumeEnabled]
-  );
-
-  useImageEffect((image) => view3d.setVolumeChannelOptions(image, index, { isosurfaceEnabled }), [isosurfaceEnabled]);
-
-  useImageEffect((image) => view3d.setVolumeChannelOptions(image, index, { isovalue }), [isovalue]);
-
-  useImageEffect((image) => view3d.setVolumeChannelOptions(image, index, { isosurfaceOpacity: opacity }), [opacity]);
-
-  useImageEffect(
-    (image) => {
-      view3d.setVolumeChannelOptions(image, index, { color });
-      view3d.updateLuts(image);
-    },
-    [color]
-  );
-
-  useImageEffect(
-    (image) => {
-      if (colorizeEnabled) {
-        // TODO get the labelColors from the tf editor component
-        const lut = image.getHistogram(index).lutGenerator_labelColors();
-        image.setColorPalette(index, lut.lut);
-        image.setColorPaletteAlpha(index, colorizeAlpha);
-      } else {
-        image.setColorPaletteAlpha(index, 0);
-      }
-      view3d.updateLuts(image);
-    },
-    [colorizeEnabled]
-  );
-
-  useImageEffect(
-    (image) => {
-      const gradient = controlPointsToLut(controlPoints);
-      image.setLut(index, gradient);
-      view3d.updateLuts(image);
-    },
-    [controlPoints]
-  );
-
-  useImageEffect(
-    (image) => {
-      image.setColorPaletteAlpha(index, colorizeEnabled ? colorizeAlpha : 0);
-      view3d.updateLuts(image);
-    },
-    [colorizeAlpha]
-  );
-
-  return null;
-};
 
 const App: React.FC<AppProps> = (props) => {
   props = { ...defaultProps, ...props };
